+++ inbox/core/app_push/backends/firebase.py
@@ -16,7 +16,11 @@ class AppPushBackend(BaseAppPushBackend):
     def __init__(self, fail_silently=False):
         super().__init__(fail_silently=fail_silently)
 
-        self.firebase_app = firebase_admin.initialize_app()
+        try:
+            self.firebase_app = firebase_admin.get_app()
+        except ValueError:
+            self.firebase_app = firebase_admin.initialize_app()
+
         self.messaging = firebase_admin.messaging
 
     def send_messages(self, messages: List[AppPushMessage]):
@@ -30,15 +34,14 @@ class AppPushBackend(BaseAppPushBackend):
             # TODO Handle failing silently if it is set to true
             self.messaging.send(message)
 
-    @property
     def notification_key(self, message: AppPushMessage):
 
-        if not self._get_registration_token:
+        if not self._get_notification_key:
             try:
-                self._get_notification_key = import_string(settings.INBOX_CONFIG.APP_PUSH_NOTIFICATION_KEY_GETTER)
+                self._get_notification_key = import_string(settings.INBOX_CONFIG['APP_PUSH_NOTIFICATION_KEY_GETTER'])
             except ImportError:
                 raise NotImplementedError('unable to load notification key getter')
 
-        notification_key = self._get_notification_key(message.entity)
+        notification_key = self._get_notification_key(message)
 
         return notification_key