+++ RELEASES.md
@@ -1,9 +1,11 @@
 # Releases
 
-### 0.3.6 (2020-05-20)
+### 0.3.13 (2020-05-20)
 
 Fixes
 
+- Fix Firebase backend
+- Don't try to initial the Firebase Admin more than once or it'll crash, so try to get it first, then init.
 - Switches EmailMessage send to use HTML instead of text content type.
 
 ### 0.3.5 (2020-05-18)